# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Management API",
        "description": "Laravel 11 Business Management API with User Model and Sanctum Authentication",
        "contact": {
            "name": "API Support",
            "email": "admin@businessapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.yourdomain.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/business/dashboard": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Get business dashboard",
                "operationId": "c5e040ac4e8e4eb00c24b62d00bd7686",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/business/vendors": {
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Create a new vendor",
                "operationId": "422fb9d58c9b9b99dda4204ef2ed859e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Supplier ABC"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "supplier@abc.com"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Office Supplies"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendor created successfully"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/api/business/purchase-orders": {
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Create purchase order",
                "operationId": "7a8308ea3c47886e073c42ece6747393",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vendor_id",
                                    "order_date"
                                ],
                                "properties": {
                                    "vendor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "order_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-15"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "draft",
                                            "pending"
                                        ],
                                        "example": "draft"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string",
                                                    "example": "Office supplies"
                                                },
                                                "quantity": {
                                                    "type": "number",
                                                    "example": 10
                                                },
                                                "unit_price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 25.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase order created"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "description": "Enter your Sanctum authentication token",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}